# Conflicts:

#	ExpressManageSystemGAP/common/src/main/java/gap/common/netconfig/RMIConfig.java

#	ExpressManageSystemGAP/server/src/main/java/gap/server/databaseutility/Config.java

#	ExpressManageSystemGAP/server/target/classes/gap/server/databaseutility/Config.class

#	ExpressManageSystemGAP/stub_driver/target/classes/gap/stub_driver/blservice/expressorder/ExpressOrderService_stub.class

#	ExpressManageSystemGAP/stub_driver/target/classes/gap/stub_driver/dataservice/expressorder/ExpressOrderDataService_stub.class

#	ExpressManageSystemGAP/stub_driver/target/classes/gap/stub_driver/dataservice/institution/InstitutionDataService_stub.class

#	ExpressManageSystemGAP/stub_driver/target/classes/gap/stub_driver/dataservice/inventory/InventoryDataService_driver.class

#	ExpressManageSystemGAP/stub_driver/target/classes/gap/stub_driver/dataservice/inventory/InventoryDataService_stub.class

#	ExpressManageSystemGAP/stub_driver/target/classes/gap/stub_driver/dataservice/loadorder/LoadOrderDataService_driver.class

#	ExpressManageSystemGAP/stub_driver/target/classes/gap/stub_driver/dataservice/loadorder/LoadOrderDataService_stub.class

#	ExpressManageSystemGAP/stub_driver/target/classes/gap/stub_driver/dataservice/pricedata/PriceDataService_driver.class

#	ExpressManageSystemGAP/stub_driver/target/classes/gap/stub_driver/dataservice/pricedata/PriceDataService_stub.class

#	ExpressManageSystemGAP/stub_driver/target/classes/gap/stub_driver/dataservice/stockindata/StockinOrderDataService_driver.class

#	ExpressManageSystemGAP/stub_driver/target/classes/gap/stub_driver/dataservice/stockindata/StockinOrderDataService_stub.class

#	ExpressManageSystemGAP/stub_driver/target/classes/gap/stub_driver/dataservice/stockoutdata/StockoutOrderDataService_driver.class

#	ExpressManageSystemGAP/stub_driver/target/classes/gap/stub_driver/dataservice/stockoutdata/StockoutOrderDataService_stub.class

#	ExpressManageSystemGAP/stub_driver/target/classes/gap/stub_driver/dataservice/userdata/UserDataService_driver.class

#	ExpressManageSystemGAP/stub_driver/target/classes/gap/stub_driver/dataservice/userdata/UserDataService_stub.class

+++ ExpressManageSystemGAP/common/src/main/java/gap/common/po/TransFarePO.java
@@ -2,7 +2,7 @@ package gap.common.po;
 
 public class TransFarePO {
 
-	private int fare;
+	private double fare;
 	private String orderID;
 	private String carID;
 
@@ -13,11 +13,11 @@ public class TransFarePO {
 
 	}
 
-	public int getFare() {
+	public double getFare() {
 		return fare;
 	}
 
-	public void setFare(int fare) {
+	public void setFare(double fare) {
 		this.fare = fare;
 	}
